+++ Frends.SFTP.DownloadFiles/Frends.SFTP.DownloadFiles/Definitions/Util.cs
@@ -1,4 +1,6 @@
 ﻿using System.Text.RegularExpressions;
+using System.Security.Cryptography;
+using System.Text;
 
 namespace Frends.SFTP.DownloadFiles.Definitions;
 
@@ -39,6 +41,82 @@ internal static class Util
     internal static byte[] ConvertFingerprintToByteArray(string fingerprint)
     {
         return fingerprint.Split(':').Select(s => Convert.ToByte(s, 16)).ToArray();
+    }
+
+    internal static string ToHex(byte[] bytes)
+    {
+        StringBuilder result = new StringBuilder(bytes.Length * 2);
+        for (int i = 0; i < bytes.Length; i++)
+            result.Append(bytes[i].ToString("x2"));
+        return result.ToString();
+    }
+
+    internal static bool TryConvertHexStringToHex(string hex)
+    {
+        try
+        {
+            var arr = new byte[hex.Length / 2];
+            for (var i = 0; i < arr.Length; i++)
+            {
+                arr[i] = Convert.ToByte(hex.Substring(i * 2, 2), 16);
+            }
+            return true;
+        } catch { return false; }
+    }
+
+    internal static string HexStringToB64String(string input)
+    {
+        return Convert.ToBase64String(ConvertHexStringToHex(input));
+    }
+
+    internal static byte[] ConvertHexStringToHex(string hex)
+    {
+        var arr = new byte[hex.Length / 2];
+        for (var i = 0; i < arr.Length; i++)
+        {
+            arr[i] = Convert.ToByte(hex.Substring(i * 2, 2), 16);
+        }
+        return arr;
+    }
+
+    internal static bool IsMD5(string input)
+    {
+        if (String.IsNullOrEmpty(input))
+        {
+            return false;
+        }
+
+        return Regex.IsMatch(input, "^[0-9a-fA-F]{32}$");
+    }
+
+    internal static bool IsSha1(string input)
+    {
+        if (String.IsNullOrEmpty(input))
+        {
+            return false;
+        }
+
+        return Regex.IsMatch(input, "^[0-9a-fA-F]{40}$");
+    }
+
+    internal static bool IsSha256(string input)
+    {
+        if (String.IsNullOrEmpty(input))
+        {
+            return false;
+        }
+
+        if (Regex.IsMatch(input, "^[0-9a-fA-F]{64}$"))
+            return true;
+
+        try
+        {
+            if (!input.EndsWith("="))
+                input += '=';
+            Convert.FromBase64String(input);
+            return true;
+        } 
+        catch { return false; }
     }
 }