+++ Frends.SFTP.DownloadFiles/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [2.0.3] - 2022-08-19
+### Fixed
+- Fixed issue with server fingerprint given by user in SHA256 hex format was not accepted: Added conversion to the fingerprint given by user.
+- Fixed issue when using invalid server fingerprint in MD5 string format throws wrong error message: Added more specific error messages.
+- Changed how MD5 string is handled. MD5 can now be given without ':' or '-' characters.
+- Fixed issue that Sha256 was only accepted in Base64 format: Added support for Sha256 in hex format.
+- Changed the used HostKeyAlgorithm by forcing to use ssh-rsa default was ed25519.
+- Added more tests for using server fingerprints.
+
 ## [2.0.2] - 2022-08-10
 ### Fixed
 - Fixed issue where when using Rename options and appending / overwrite the task would throw an exception because the work directory given by user was used with renaming destination file.